use double

+++ include/mu/literals.h
@@ -5,9 +5,9 @@
 
 namespace mu {
 
-constexpr float PI = M_PI;
-constexpr float PI2 = M_PI_2;
-constexpr float PI4 = M_PI_4;
+constexpr double PI = M_PI;
+constexpr double PI2 = M_PI_2;
+constexpr double PI4 = M_PI_4;
 
 }  // namespace mu
